tesseract_nanobind_benchmark/pyproject.toml
[build-system]
requires = ["scikit-build-core", "nanobind"]
build-backend = "scikit_build_core.build"

[project]
name = "tesseract_nanobind"
version = "0.1.0"
description = "High-performance Tesseract OCR binding using nanobind"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0",
    "pillow>=12.0",
]
benchmark = [
    "pytesseract>=0.3.10",
]

[tool.scikit-build]
cmake.version = ">=3.15"
wheel.packages = ["src/tesseract_nanobind"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
